+++ lib/array/isInstance/method.ts
@@ -1,3 +1,0 @@
-const method = Array.isArray;
-
-export = method;