+++ lib/date/isDate/test.ts
@@ -1,0 +1,13 @@
+import "./index";
+
+describe("Date.isDate", () => {
+  test("Date.isDate(\"foo bar\") returns false", () => {
+    expect(Date.isDate("foo bar"))
+      .toBe(false);
+  });
+
+  test("Date.isDate(new Date()) returns true", () => {
+    expect(Date.isDate(new Date()))
+      .toBe(true);
+  });
+});