+++ lib/function/methods.ts
@@ -1,11 +1,11 @@
 import * as cache from "./cache/method";
 import * as defer from "./defer/method";
-import * as isInstance from "./isInstance/method";
+import * as isFunction from "./isFunction/method";
 import * as once from "./once/method";
 
 export {
   cache,
   defer,
-  isInstance,
+  isFunction,
   once,
 };